Updated AdminX config boilerplate to be more reusable
