+++ apps/admin-x-design-system/postcss.config.cjs
@@ -1,4 +1,4 @@
-export default {
+module.exports = {
     plugins: {
         'postcss-import': {},
         'tailwindcss/nesting': {},
