+++ apps/admin-x-framework/package.json
@@ -29,6 +29,10 @@
     "./api/*": {
         "import": "./es/api/*.js",
         "types": "./types/api/*.d.ts"
+    },
+    "./vite": {
+        "import": "./es/vite.js",
+        "types": "./types/vite.d.ts"
     }
   },
   "sideEffects": false,
@@ -46,22 +50,27 @@
     "types"
   ],
   "devDependencies": {
-    "@vitejs/plugin-react": "4.1.1",
     "c8": "8.0.1",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-react-refresh": "0.4.3",
     "mocha": "10.2.0",
     "sinon": "17.0.0",
     "ts-node": "10.9.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "typescript": "5.2.2",
-    "vite": "4.5.0"
+    "typescript": "5.2.2"
   },
   "dependencies": {
     "@sentry/react": "7.80.1",
     "@tanstack/react-query": "4.36.1",
-    "@tryghost/admin-x-design-system": "0.0.0"
+    "@tryghost/admin-x-design-system": "0.0.0",
+    "@types/react": "18.2.37",
+    "@types/react-dom": "18.2.15",
+    "@vitejs/plugin-react": "4.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "vite": "4.5.0",
+    "vite-plugin-css-injected-by-js": "^3.3.0",
+    "vite-plugin-svgr": "3.3.0",
+    "vitest": "0.34.3"
   },
   "peerDependencies": {
     "react": "^18.2.0",
