+++ apps/admin-x-framework/vite.config.ts
@@ -10,10 +10,6 @@ export default (function viteConfig() {
         plugins: [
             react()
         ],
-        define: {
-            'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
-            'process.env.VITEST_SEGFAULT_RETRY': 3
-        },
         preview: {
             port: 4174
         },
